Destroy: delete user before deleting files
Otherwise this error happens in macOS:

```
Removing /Users/pi...
Enter sudo password (required to destroy installation):
Error: Error: sudo command 'rm -rf /Users/pi' failed (exit code 1): rm: /Users/pi/Pictures: Operation not permitted
rm: /Users/pi/Library/Application Support: Operation not permitted
rm: /Users/pi/Library/Preferences: Operation not permitted
rm: /Users/pi/Library/Mail: Operation not permitted
rm: /Users/pi/Library/Containers/com.apple.ImageIO.imageimporter/.com.apple.containermanagerd.metadata.plist: Operation not permitted
rm: /Users/pi/Library/Containers/com.apple.ImageIO.imageimporter: Operation not permitted
rm: /Users/pi/Library/Containers/com.apple.PDFKit.PDFImporter/.com.apple.containermanagerd.metadata.plist: Operation not permitted
rm: /Users/pi/Library/Containers/com.apple.PDFKit.PDFImporter: Operation not permitted
rm: /Users/pi/Library/Containers: Operation not permitted
rm: /Users/pi/Library/ContainerManager: Operation not permitted
rm: /Users/pi/Library/Caches/com.apple.containermanagerd: Operation not permitted
rm: /Users/pi/Library/Caches: Operation not permitted
rm: /Users/pi/Library: Operation not permitted
rm: /Users/pi: Operation not permitted
    at ChildProcess.<anonymous> (/Users/pi/Work/skynot/dist/index.js:149:24)
    at ChildProcess.emit (node:events:509:20)
    at maybeClose (node:internal/child_process:1108:16)
    at ChildProcess._handle.onexit (node:internal/child_process:305:5)

```

+++ src/index.ts
@@ -532,19 +532,23 @@ async function destroyInstallation(): Promise<void> {
 
   const reason = 'required to destroy installation';
 
-  // Remove the user's home directory (all data)
-  console.log(`Removing ${piHome}...`);
-  await askSudoPasswordAndRun(`rm -rf ${piHome}`, reason);
-  console.log('User data deleted.');
-
-  // Delete the user
+  // Delete the user first (which also removes the home directory on Linux with -r, and on macOS sysadminctl removes the home)
   console.log(`Deleting user '${AGENT_USER}'...`);
   if (platform === 'darwin') {
+    // sysadminctl deletes the user and its home directory by default
     await askSudoPasswordAndRun(`sysadminctl -deleteUser ${AGENT_USER}`, reason);
   } else {
-    await askSudoPasswordAndRun(`userdel ${AGENT_USER}`, reason);
+    // -r flag removes the home directory
+    await askSudoPasswordAndRun(`userdel -r ${AGENT_USER}`, reason);
   }
-  console.log('User deleted.');
+  console.log('User deleted (home directory removed).');
+
+  // Ensure home directory is gone (some macOS configs may leave it)
+  if (fs.existsSync(piHome)) {
+    console.log(`Cleaning residual home directory ${piHome}...`);
+    await askSudoPasswordAndRun(`rm -rf ${piHome}`, reason);
+    console.log('Residual home directory removed.');
+  }
 
   // Delete the group
   console.log(`Deleting group '${AGENT_GROUP_NAME}'...`);
